% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculatePercentGenes.R
\name{calculatePercentGenes}
\alias{calculatePercentGenes}
\title{Calculate percentage of transcripts of gene list.}
\usage{
calculatePercentGenes(object, assay = "RNA", genes)
}
\arguments{
\item{object}{Seurat object.}

\item{assay}{Assay to pull counts from; defaults to 'RNA'. Only relevant in
Seurat v3.0 or higher since the concept of assays wasn't implemented before.}

\item{genes}{List(s) of genes.}
}
\value{
List of lists containing the percentages of expression for each
provided gene list.
}
\description{
Get percentage of transcripts of gene list compared to all
transcripts per cell.
}
\details{
Calculate percentage of transcripts of gene list.
}
\examples{
pbmc <- readRDS(system.file("extdata/v1.3/pbmc_seurat.rds",
  package = "cerebroApp"))
pbmc <- calculatePercentGenes(
  object = pbmc,
  assay = 'RNA',
  genes = list('example' = c('FCN1','CD3D'))
)
}
\keyword{Cerebro}
\keyword{Seurat}
\keyword{scRNAseq}
