% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cerebro_v1.3-class.R
\name{Cerebro_v1.3}
\alias{Cerebro_v1.3}
\title{R6 class in which data sets will be stored for visualization in Cerebro.}
\value{
A new \code{Cerebro_v1.3} object.
}
\description{
A \code{Cerebro_v1.3} object is an R6 class that contains several types of
data that can be visualized in Cerebro.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{version}}{cerebroApp version that was used to create the object.}

\item{\code{experiment}}{\code{list} that contains meta data about the data set,
including experiment name, species, date of export.}

\item{\code{technical_info}}{\code{list} that contains technical information
about the analysis, including the R session info.}

\item{\code{parameters}}{\code{list} that contains important parameters that
were used during the analysis, e.g. cut-off values for cell filtering.}

\item{\code{groups}}{\code{list} that contains specified grouping variables and
and the group levels (subgroups) that belong to each of them. For each
grouping variable, a corresponding column with the same name must exist
in the meta data.}

\item{\code{cell_cycle}}{\code{vector} that contains the name of columns in the
meta data that contain cell cycle assignments.}

\item{\code{gene_lists}}{\code{list} that contains gene lists, e.g.
mitochondrial and/or ribosomal genes.}

\item{\code{expression}}{\code{matrix}-like object that holds transcript counts.}

\item{\code{meta_data}}{\code{data.frame} that contains cell meta data.}

\item{\code{projections}}{\code{list} that contains projections/dimensional
reductions.}

\item{\code{most_expressed_genes}}{\code{list} that contains a \code{data.frame}
holding the most expressed genes for each grouping variable that was
specified during the call to \code{\link{getMostExpressedGenes}}.}

\item{\code{marker_genes}}{\code{list} that contains a \code{list} for every
method that was used to calculate marker genes, and a \code{data.frame}
for each grouping variable, e.g. those that were specified during the
call to \code{\link{getMarkerGenes}}.}

\item{\code{enriched_pathways}}{\code{list} that contains a \code{list} for
every method that was used to calculate marker genes, and a
\code{data.frame} for each grouping variable, e.g. those that were
specified during the call to \code{\link{getEnrichedPathways}} or
\code{\link{performGeneSetEnrichmentAnalysis}}.}

\item{\code{trees}}{\code{list} that contains a phylogenetic tree (class
\code{phylo}) for grouping variables.}

\item{\code{trajectories}}{\code{list} that contains a \code{list} for every
method that was used to calculate trajectories, and, depending on the
method, a \code{data.frame} or \code{list} for each specific trajectory,
e.g. those extracted with \code{\link{extractMonocleTrajectory}}.}

\item{\code{extra_material}}{\code{list} that can contain additional material
related to the data set; tables should be stored in \code{data.frame}
format in a named \code{list} called `tables`}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Cerebro_v1.3$new()}}
\item \href{#method-setVersion}{\code{Cerebro_v1.3$setVersion()}}
\item \href{#method-getVersion}{\code{Cerebro_v1.3$getVersion()}}
\item \href{#method-checkIfGroupExists}{\code{Cerebro_v1.3$checkIfGroupExists()}}
\item \href{#method-checkIfColumnExistsInMetadata}{\code{Cerebro_v1.3$checkIfColumnExistsInMetadata()}}
\item \href{#method-addExperiment}{\code{Cerebro_v1.3$addExperiment()}}
\item \href{#method-getExperiment}{\code{Cerebro_v1.3$getExperiment()}}
\item \href{#method-addParameters}{\code{Cerebro_v1.3$addParameters()}}
\item \href{#method-getParameters}{\code{Cerebro_v1.3$getParameters()}}
\item \href{#method-addTechnicalInfo}{\code{Cerebro_v1.3$addTechnicalInfo()}}
\item \href{#method-getTechnicalInfo}{\code{Cerebro_v1.3$getTechnicalInfo()}}
\item \href{#method-addGroup}{\code{Cerebro_v1.3$addGroup()}}
\item \href{#method-getGroups}{\code{Cerebro_v1.3$getGroups()}}
\item \href{#method-getGroupLevels}{\code{Cerebro_v1.3$getGroupLevels()}}
\item \href{#method-setMetaData}{\code{Cerebro_v1.3$setMetaData()}}
\item \href{#method-getMetaData}{\code{Cerebro_v1.3$getMetaData()}}
\item \href{#method-addGeneList}{\code{Cerebro_v1.3$addGeneList()}}
\item \href{#method-getGeneLists}{\code{Cerebro_v1.3$getGeneLists()}}
\item \href{#method-setExpression}{\code{Cerebro_v1.3$setExpression()}}
\item \href{#method-getCellNames}{\code{Cerebro_v1.3$getCellNames()}}
\item \href{#method-getGeneNames}{\code{Cerebro_v1.3$getGeneNames()}}
\item \href{#method-getMeanExpressionForGenes}{\code{Cerebro_v1.3$getMeanExpressionForGenes()}}
\item \href{#method-getMeanExpressionForCells}{\code{Cerebro_v1.3$getMeanExpressionForCells()}}
\item \href{#method-getExpressionMatrix}{\code{Cerebro_v1.3$getExpressionMatrix()}}
\item \href{#method-setCellCycle}{\code{Cerebro_v1.3$setCellCycle()}}
\item \href{#method-getCellCycle}{\code{Cerebro_v1.3$getCellCycle()}}
\item \href{#method-addProjection}{\code{Cerebro_v1.3$addProjection()}}
\item \href{#method-availableProjections}{\code{Cerebro_v1.3$availableProjections()}}
\item \href{#method-getProjection}{\code{Cerebro_v1.3$getProjection()}}
\item \href{#method-addTree}{\code{Cerebro_v1.3$addTree()}}
\item \href{#method-getTree}{\code{Cerebro_v1.3$getTree()}}
\item \href{#method-addMostExpressedGenes}{\code{Cerebro_v1.3$addMostExpressedGenes()}}
\item \href{#method-getGroupsWithMostExpressedGenes}{\code{Cerebro_v1.3$getGroupsWithMostExpressedGenes()}}
\item \href{#method-getMostExpressedGenes}{\code{Cerebro_v1.3$getMostExpressedGenes()}}
\item \href{#method-addMarkerGenes}{\code{Cerebro_v1.3$addMarkerGenes()}}
\item \href{#method-getMethodsForMarkerGenes}{\code{Cerebro_v1.3$getMethodsForMarkerGenes()}}
\item \href{#method-getGroupsWithMarkerGenes}{\code{Cerebro_v1.3$getGroupsWithMarkerGenes()}}
\item \href{#method-getMarkerGenes}{\code{Cerebro_v1.3$getMarkerGenes()}}
\item \href{#method-addEnrichedPathways}{\code{Cerebro_v1.3$addEnrichedPathways()}}
\item \href{#method-getMethodsForEnrichedPathways}{\code{Cerebro_v1.3$getMethodsForEnrichedPathways()}}
\item \href{#method-getGroupsWithEnrichedPathways}{\code{Cerebro_v1.3$getGroupsWithEnrichedPathways()}}
\item \href{#method-getEnrichedPathways}{\code{Cerebro_v1.3$getEnrichedPathways()}}
\item \href{#method-addTrajectory}{\code{Cerebro_v1.3$addTrajectory()}}
\item \href{#method-getMethodsForTrajectories}{\code{Cerebro_v1.3$getMethodsForTrajectories()}}
\item \href{#method-getNamesOfTrajectories}{\code{Cerebro_v1.3$getNamesOfTrajectories()}}
\item \href{#method-getTrajectory}{\code{Cerebro_v1.3$getTrajectory()}}
\item \href{#method-addExtraMaterial}{\code{Cerebro_v1.3$addExtraMaterial()}}
\item \href{#method-addExtraTable}{\code{Cerebro_v1.3$addExtraTable()}}
\item \href{#method-getExtraMaterialCategories}{\code{Cerebro_v1.3$getExtraMaterialCategories()}}
\item \href{#method-checkForExtraTables}{\code{Cerebro_v1.3$checkForExtraTables()}}
\item \href{#method-getNamesOfExtraTables}{\code{Cerebro_v1.3$getNamesOfExtraTables()}}
\item \href{#method-getExtraTable}{\code{Cerebro_v1.3$getExtraTable()}}
\item \href{#method-print}{\code{Cerebro_v1.3$print()}}
\item \href{#method-clone}{\code{Cerebro_v1.3$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new \code{Cerebro_v1.3} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cerebro_v1.3$new()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A new \code{Cerebro_v1.3} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-setVersion"></a>}}
\if{latex}{\out{\hypertarget{method-setVersion}{}}}
\subsection{Method \code{setVersion()}}{
Set the version of \code{cerebroApp} that was used to generate this
object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cerebro_v1.3$setVersion(version)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{version}}{Version to set.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getVersion"></a>}}
\if{latex}{\out{\hypertarget{method-getVersion}{}}}
\subsection{Method \code{getVersion()}}{
Get the version of \code{cerebroApp} that was used to generate this
object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cerebro_v1.3$getVersion()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Version as \code{package_version} class.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-checkIfGroupExists"></a>}}
\if{latex}{\out{\hypertarget{method-checkIfGroupExists}{}}}
\subsection{Method \code{checkIfGroupExists()}}{
Safety function that will check if a provided group name is present in
the \code{groups} field.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cerebro_v1.3$checkIfGroupExists(group_name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{group_name}}{Group name to be tested}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-checkIfColumnExistsInMetadata"></a>}}
\if{latex}{\out{\hypertarget{method-checkIfColumnExistsInMetadata}{}}}
\subsection{Method \code{checkIfColumnExistsInMetadata()}}{
Safety function that will check if a provided group name is present in
the meta data.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cerebro_v1.3$checkIfColumnExistsInMetadata(group_name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{group_name}}{Group name to be tested.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-addExperiment"></a>}}
\if{latex}{\out{\hypertarget{method-addExperiment}{}}}
\subsection{Method \code{addExperiment()}}{
Add information to \code{experiment} field.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cerebro_v1.3$addExperiment(field, content)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{field}}{Name of the information, e.g. \code{organism}.}

\item{\code{content}}{Actual information, e.g. \code{hg}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getExperiment"></a>}}
\if{latex}{\out{\hypertarget{method-getExperiment}{}}}
\subsection{Method \code{getExperiment()}}{
Retrieve information from \code{experiment} field.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cerebro_v1.3$getExperiment()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{list} of all entries in the \code{experiment} field.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-addParameters"></a>}}
\if{latex}{\out{\hypertarget{method-addParameters}{}}}
\subsection{Method \code{addParameters()}}{
Add information to \code{parameters} field.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cerebro_v1.3$addParameters(field, content)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{field}}{Name of the information, e.g. \code{number_of_PCs}.}

\item{\code{content}}{Actual information, e.g. \code{30}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getParameters"></a>}}
\if{latex}{\out{\hypertarget{method-getParameters}{}}}
\subsection{Method \code{getParameters()}}{
Retrieve information from \code{parameters} field.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cerebro_v1.3$getParameters()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{list} of all entries in the \code{parameters} field.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-addTechnicalInfo"></a>}}
\if{latex}{\out{\hypertarget{method-addTechnicalInfo}{}}}
\subsection{Method \code{addTechnicalInfo()}}{
Add information to \code{technical_info} field.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cerebro_v1.3$addTechnicalInfo(field, content)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{field}}{Name of the information, e.g. \code{R}.}

\item{\code{content}}{Actual information, e.g. \code{4.0.2}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getTechnicalInfo"></a>}}
\if{latex}{\out{\hypertarget{method-getTechnicalInfo}{}}}
\subsection{Method \code{getTechnicalInfo()}}{
Retrieve information from \code{technical_info} field.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cerebro_v1.3$getTechnicalInfo()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{list} of all entries in the \code{technical_info} field.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-addGroup"></a>}}
\if{latex}{\out{\hypertarget{method-addGroup}{}}}
\subsection{Method \code{addGroup()}}{
Add group to the groups registered in the \code{groups} field.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cerebro_v1.3$addGroup(group_name, levels)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{group_name}}{Group name.}

\item{\code{levels}}{\code{vector} of group levels (subgroups).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getGroups"></a>}}
\if{latex}{\out{\hypertarget{method-getGroups}{}}}
\subsection{Method \code{getGroups()}}{
Retrieve all names in the \code{groups} field.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cerebro_v1.3$getGroups()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{vector} of registered groups.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getGroupLevels"></a>}}
\if{latex}{\out{\hypertarget{method-getGroupLevels}{}}}
\subsection{Method \code{getGroupLevels()}}{
Retrieve group levels for a group registered in the \code{groups} field.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cerebro_v1.3$getGroupLevels(group_name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{group_name}}{Group name for which to retrieve group levels.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{vector} of group levels.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-setMetaData"></a>}}
\if{latex}{\out{\hypertarget{method-setMetaData}{}}}
\subsection{Method \code{setMetaData()}}{
Set meta data for cells.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cerebro_v1.3$setMetaData(table)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{table}}{\code{data.frame} that contains meta data for cells. The
number of rows must be equal to the number of rows of projections and
the number of columns in the transcript count matrix.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getMetaData"></a>}}
\if{latex}{\out{\hypertarget{method-getMetaData}{}}}
\subsection{Method \code{getMetaData()}}{
Retrieve meta data for cells.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cerebro_v1.3$getMetaData()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{data.frame} containing meta data.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-addGeneList"></a>}}
\if{latex}{\out{\hypertarget{method-addGeneList}{}}}
\subsection{Method \code{addGeneList()}}{
Add a gene list to the \code{gene_lists}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cerebro_v1.3$addGeneList(name, genes)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{Name of the gene list.}

\item{\code{genes}}{\code{vector} of genes.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getGeneLists"></a>}}
\if{latex}{\out{\hypertarget{method-getGeneLists}{}}}
\subsection{Method \code{getGeneLists()}}{
Retrieve gene lists from the \code{gene_lists}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cerebro_v1.3$getGeneLists()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{list} of all entries in the \code{gene_lists} field.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-setExpression"></a>}}
\if{latex}{\out{\hypertarget{method-setExpression}{}}}
\subsection{Method \code{setExpression()}}{
Set transcript count matrix.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cerebro_v1.3$setExpression(counts)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{counts}}{\code{matrix}-like object that contains transcript counts
for cells in the data set. Number of columns must be equal to the number
of rows in the \code{meta_data} field.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getCellNames"></a>}}
\if{latex}{\out{\hypertarget{method-getCellNames}{}}}
\subsection{Method \code{getCellNames()}}{
Get names of all cells.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cerebro_v1.3$getCellNames()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{vector} containing all cell names/barcodes.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getGeneNames"></a>}}
\if{latex}{\out{\hypertarget{method-getGeneNames}{}}}
\subsection{Method \code{getGeneNames()}}{
Get names of all genes in transcript count matrix.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cerebro_v1.3$getGeneNames()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{vector} containing all gene names in transcript count matrix.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getMeanExpressionForGenes"></a>}}
\if{latex}{\out{\hypertarget{method-getMeanExpressionForGenes}{}}}
\subsection{Method \code{getMeanExpressionForGenes()}}{
Retrieve mean expression across all cells in the data set for a set of
genes.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cerebro_v1.3$getMeanExpressionForGenes(genes)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{genes}}{Names of genes to extract; no default.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{data.frame} containing specified gene names and their respective
mean expression across all cells in the data set.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getMeanExpressionForCells"></a>}}
\if{latex}{\out{\hypertarget{method-getMeanExpressionForCells}{}}}
\subsection{Method \code{getMeanExpressionForCells()}}{
Retrieve (mean) expression for a single gene or a set of genes for a
given set of cells.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cerebro_v1.3$getMeanExpressionForCells(cells = NULL, genes = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{cells}}{Names/barcodes of cells to extract; defaults to \code{NULL},
which will return all cells.}

\item{\code{genes}}{Names of genes to extract; defaults to \code{NULL}, which
will return all genes.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{vector} containing (mean) expression across all specified genes in
each specified cell.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getExpressionMatrix"></a>}}
\if{latex}{\out{\hypertarget{method-getExpressionMatrix}{}}}
\subsection{Method \code{getExpressionMatrix()}}{
Retrieve transcript count matrix.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cerebro_v1.3$getExpressionMatrix(cells = NULL, genes = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{cells}}{Names/barcodes of cells to extract; defaults to \code{NULL},
which will return all cells.}

\item{\code{genes}}{Names of genes to extract; defaults to \code{NULL}, which
will return all genes.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Dense transcript count matrix for specified cells and genes.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-setCellCycle"></a>}}
\if{latex}{\out{\hypertarget{method-setCellCycle}{}}}
\subsection{Method \code{setCellCycle()}}{
Add columns containing cell cycle assignments to the \code{cell_cycle}
field.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cerebro_v1.3$setCellCycle(cols)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{cols}}{\code{vector} of columns names containing cell cycle
assignments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getCellCycle"></a>}}
\if{latex}{\out{\hypertarget{method-getCellCycle}{}}}
\subsection{Method \code{getCellCycle()}}{
Retrieve column names containing cell cycle assignments.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cerebro_v1.3$getCellCycle()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{vector} of column names in meta data.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-addProjection"></a>}}
\if{latex}{\out{\hypertarget{method-addProjection}{}}}
\subsection{Method \code{addProjection()}}{
Add projections (dimensional reductions).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cerebro_v1.3$addProjection(name, projection)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{Name of the projection.}

\item{\code{projection}}{\code{data.frame} containing positions of cells in
projection.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-availableProjections"></a>}}
\if{latex}{\out{\hypertarget{method-availableProjections}{}}}
\subsection{Method \code{availableProjections()}}{
Get list of available projections (dimensional reductions).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cerebro_v1.3$availableProjections()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{vector} of projections / dimensional reductions that are available.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getProjection"></a>}}
\if{latex}{\out{\hypertarget{method-getProjection}{}}}
\subsection{Method \code{getProjection()}}{
Retrieve data for a specific projection.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cerebro_v1.3$getProjection(name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{Name of projection.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{data.frame} containing the positions of cells in the projection.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-addTree"></a>}}
\if{latex}{\out{\hypertarget{method-addTree}{}}}
\subsection{Method \code{addTree()}}{
Add phylogenetic tree to \code{trees} field.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cerebro_v1.3$addTree(group_name, tree)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{group_name}}{Group name that this tree belongs to.}

\item{\code{tree}}{Phylogenetic tree as \code{phylo} object.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getTree"></a>}}
\if{latex}{\out{\hypertarget{method-getTree}{}}}
\subsection{Method \code{getTree()}}{
Retrieve phylogenetic tree for a specific group.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cerebro_v1.3$getTree(group_name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{group_name}}{Group name for which to retrieve phylogenetic tree.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Phylogenetic tree as \code{phylo} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-addMostExpressedGenes"></a>}}
\if{latex}{\out{\hypertarget{method-addMostExpressedGenes}{}}}
\subsection{Method \code{addMostExpressedGenes()}}{
Add table of most expressed genes.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cerebro_v1.3$addMostExpressedGenes(group_name, table)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{group_name}}{Name of grouping variable that the most expressed genes
belong to. Must be registered in the \code{groups} field.}

\item{\code{table}}{\code{data.frame} that contains the most expressed genes.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getGroupsWithMostExpressedGenes"></a>}}
\if{latex}{\out{\hypertarget{method-getGroupsWithMostExpressedGenes}{}}}
\subsection{Method \code{getGroupsWithMostExpressedGenes()}}{
Retrieve names of grouping variables for which most expressed genes are
available.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cerebro_v1.3$getGroupsWithMostExpressedGenes()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{vector} of grouping variables for which most expressed genes are
available.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getMostExpressedGenes"></a>}}
\if{latex}{\out{\hypertarget{method-getMostExpressedGenes}{}}}
\subsection{Method \code{getMostExpressedGenes()}}{
Retrieve table of most expressed genes for a grouping variable.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cerebro_v1.3$getMostExpressedGenes(group_name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{group_name}}{Grouping variable for which most expressed genes should
be retrieved.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{data.frame} that contains most expressed genes for group levels of
the specified grouping variable.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-addMarkerGenes"></a>}}
\if{latex}{\out{\hypertarget{method-addMarkerGenes}{}}}
\subsection{Method \code{addMarkerGenes()}}{
Add table of marker genes.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cerebro_v1.3$addMarkerGenes(method, name, table)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{method}}{Name of method that was used to generate the marker genes.}

\item{\code{name}}{Name of table. This name will be used to select the table in
Cerebro. It is recommended to use the grouping variable, e.g.
\code{sample}.}

\item{\code{table}}{\code{data.frame} that contains the marker genes.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getMethodsForMarkerGenes"></a>}}
\if{latex}{\out{\hypertarget{method-getMethodsForMarkerGenes}{}}}
\subsection{Method \code{getMethodsForMarkerGenes()}}{
Retrieve names of methods that were used to generate marker genes.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cerebro_v1.3$getMethodsForMarkerGenes()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{vector} of names of methods that were used to generate marker
genes.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getGroupsWithMarkerGenes"></a>}}
\if{latex}{\out{\hypertarget{method-getGroupsWithMarkerGenes}{}}}
\subsection{Method \code{getGroupsWithMarkerGenes()}}{
Retrieve grouping variables for which marker genes were generated using
a specified method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cerebro_v1.3$getGroupsWithMarkerGenes(method)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{method}}{Name of method.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{vector} of grouping variables for which marker genes were
calculated using the specified method.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getMarkerGenes"></a>}}
\if{latex}{\out{\hypertarget{method-getMarkerGenes}{}}}
\subsection{Method \code{getMarkerGenes()}}{
Retrieve table of marker genes for specific method and grouping variable.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cerebro_v1.3$getMarkerGenes(method, name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{method}}{Name of method.}

\item{\code{name}}{Name of table.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{data.frame} that contains marker genes for the specified
combination of method and grouping variable.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-addEnrichedPathways"></a>}}
\if{latex}{\out{\hypertarget{method-addEnrichedPathways}{}}}
\subsection{Method \code{addEnrichedPathways()}}{
Add table of enriched pathways.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cerebro_v1.3$addEnrichedPathways(method, name, table)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{method}}{Name of method that was used to generate the enriched
pathways.}

\item{\code{name}}{Name of table. This name will be used to select the table in
Cerebro. It is recommended to use the grouping variable, e.g.
\code{sample}.}

\item{\code{table}}{\code{data.frame} that contains the enriched pathways.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getMethodsForEnrichedPathways"></a>}}
\if{latex}{\out{\hypertarget{method-getMethodsForEnrichedPathways}{}}}
\subsection{Method \code{getMethodsForEnrichedPathways()}}{
Retrieve names of methods that were used to generate enriched pathways.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cerebro_v1.3$getMethodsForEnrichedPathways()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{vector} of names of methods that were used to generate enriched
pathways.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getGroupsWithEnrichedPathways"></a>}}
\if{latex}{\out{\hypertarget{method-getGroupsWithEnrichedPathways}{}}}
\subsection{Method \code{getGroupsWithEnrichedPathways()}}{
Retrieve grouping variables for which enriched pathways were generated
using a specified method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cerebro_v1.3$getGroupsWithEnrichedPathways(method)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{method}}{Name of method.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{vector} of grouping variables for which enriched pathways were
calculated using the specified method.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getEnrichedPathways"></a>}}
\if{latex}{\out{\hypertarget{method-getEnrichedPathways}{}}}
\subsection{Method \code{getEnrichedPathways()}}{
Retrieve table of enriched pathways for specific method and grouping
variable.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cerebro_v1.3$getEnrichedPathways(method, name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{method}}{Name of method.}

\item{\code{name}}{Grouping variable.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{data.frame} that contains enriched pathways for the specified
combination of method and grouping variable.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-addTrajectory"></a>}}
\if{latex}{\out{\hypertarget{method-addTrajectory}{}}}
\subsection{Method \code{addTrajectory()}}{
Add trajectory.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cerebro_v1.3$addTrajectory(method, name, content)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{method}}{Name of method that was used to generate the trajectory.}

\item{\code{name}}{Name of the trajectory. This name will be used later in
Cerebro to select the trajectory.}

\item{\code{content}}{Relevant data for the trajectory, depending on the method
this could be a \code{list} holding edges, cell positions, pseudotime,
etc.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getMethodsForTrajectories"></a>}}
\if{latex}{\out{\hypertarget{method-getMethodsForTrajectories}{}}}
\subsection{Method \code{getMethodsForTrajectories()}}{
Retrieve names of methods that were used to generate trajectories.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cerebro_v1.3$getMethodsForTrajectories()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{vector} of names of methods that were used to generate
trajectories.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getNamesOfTrajectories"></a>}}
\if{latex}{\out{\hypertarget{method-getNamesOfTrajectories}{}}}
\subsection{Method \code{getNamesOfTrajectories()}}{
Retrieve names of available trajectories for a specified method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cerebro_v1.3$getNamesOfTrajectories(method)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{method}}{Name of method.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{vector} of available trajectory for the specified method.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getTrajectory"></a>}}
\if{latex}{\out{\hypertarget{method-getTrajectory}{}}}
\subsection{Method \code{getTrajectory()}}{
Retrieve data for a specific trajectory.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cerebro_v1.3$getTrajectory(method, name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{method}}{Name of method.}

\item{\code{name}}{Name of trajectory.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The type of data depends on the method that was used to generate the
trajectory.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-addExtraMaterial"></a>}}
\if{latex}{\out{\hypertarget{method-addExtraMaterial}{}}}
\subsection{Method \code{addExtraMaterial()}}{
Add content to extra material field.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cerebro_v1.3$addExtraMaterial(category, name, content)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{category}}{Name of category. At the moment, only \code{tables} is a
valid category.}

\item{\code{name}}{Name of material, will be used to select it in Cerebro.}

\item{\code{content}}{Data that should be added.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-addExtraTable"></a>}}
\if{latex}{\out{\hypertarget{method-addExtraTable}{}}}
\subsection{Method \code{addExtraTable()}}{
Add table to `extra_material` slot.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cerebro_v1.3$addExtraTable(name, table)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{Name of material, will be used to select it in Cerebro.}

\item{\code{table}}{Table that should be added, must be \code{data.frame}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getExtraMaterialCategories"></a>}}
\if{latex}{\out{\hypertarget{method-getExtraMaterialCategories}{}}}
\subsection{Method \code{getExtraMaterialCategories()}}{
Get names of categories for which extra material is available.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cerebro_v1.3$getExtraMaterialCategories()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{vector} with names of available categories.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-checkForExtraTables"></a>}}
\if{latex}{\out{\hypertarget{method-checkForExtraTables}{}}}
\subsection{Method \code{checkForExtraTables()}}{
Check whether there are tables in the extra materials.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cerebro_v1.3$checkForExtraTables()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{logical} indicating whether there are tables in the extra
materials.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getNamesOfExtraTables"></a>}}
\if{latex}{\out{\hypertarget{method-getNamesOfExtraTables}{}}}
\subsection{Method \code{getNamesOfExtraTables()}}{
Get names of tables in extra materials.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cerebro_v1.3$getNamesOfExtraTables()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{vector} containing names of tables in extra materials.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getExtraTable"></a>}}
\if{latex}{\out{\hypertarget{method-getExtraTable}{}}}
\subsection{Method \code{getExtraTable()}}{
Get table from extra materials.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cerebro_v1.3$getExtraTable(name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{Name of table.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Requested table in \code{data.frame} format.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
Show overview of object and the data it contains.
Print overview of available marker gene results for \code{self$print()}
function.
Print overview of available enriched pathway results for
\code{self$print()} function.
Print overview of available trajectories for \code{self$print()} function.
Print overview of extra material for \code{self$print()} function.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cerebro_v1.3$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cerebro_v1.3$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
