% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_calculateTableAB.R
\name{.calculateTableAB}
\alias{.calculateTableAB}
\title{test}
\usage{
.calculateTableAB(table, groupA, groupB, mode, percent)
}
\arguments{
\item{table}{test}

\item{groupA}{test}

\item{groupB}{test}

\item{mode}{test}

\item{percent}{test}
}
\value{
test
}
\description{
Function to calculate A-by-B tables, e.g. samples by clusters.
}
