% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_centerOfGroups.R
\name{.centerOfGroups}
\alias{.centerOfGroups}
\title{test}
\usage{
.centerOfGroups(coordinates, df, n_dimensions, group)
}
\arguments{
\item{table}{test}

\item{groupA}{test}

\item{groupB}{test}

\item{mode}{test}

\item{percent}{test}
}
\value{
test
}
\description{
Function to calculate center of groups in projections/
trajectories.
}
