% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_checkMatrixFormat.R
\name{.checkMatrixFormat}
\alias{.checkMatrixFormat}
\title{test}
\usage{
.checkMatrixFormat(matrix)
}
\arguments{
\item{table}{test}

\item{groupA}{test}

\item{groupB}{test}

\item{mode}{test}

\item{percent}{test}
}
\value{
test
}
\description{
test
}
