% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_fetchExpression.R
\name{.fetchExpression}
\alias{.fetchExpression}
\title{test}
\usage{
.fetchExpression(matrix = NULL, API = NULL, genes = NULL)
}
\arguments{
\item{table}{test}

\item{groupA}{test}

\item{groupB}{test}

\item{mode}{test}

\item{percent}{test}
}
\value{
test
}
\description{
test
}
