% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_fetchExpressionFromH5File.R
\name{.fetchExpressionFromH5File}
\alias{.fetchExpressionFromH5File}
\title{test}
\usage{
.fetchExpressionFromH5File(matrix, genes = c())
}
\arguments{
\item{table}{test}

\item{groupA}{test}

\item{groupB}{test}

\item{mode}{test}

\item{percent}{test}
}
\value{
test
}
\description{
test
}
