% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_fetchExpressionFromMatrix.R
\name{.fetchExpressionFromMatrix}
\alias{.fetchExpressionFromMatrix}
\title{test}
\usage{
.fetchExpressionFromMatrix(matrix, genes = c())
}
\arguments{
\item{table}{test}

\item{groupA}{test}

\item{groupB}{test}

\item{mode}{test}

\item{percent}{test}
}
\value{
test
}
\description{
test
}
