% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_getXYranges.R
\name{.getXYranges}
\alias{.getXYranges}
\title{test}
\usage{
.getXYranges(df)
}
\arguments{
\item{table}{test}

\item{groupA}{test}

\item{groupB}{test}

\item{mode}{test}

\item{percent}{test}
}
\value{
test
}
\description{
Calculate X-Y ranges for projections.
}
