% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_prettifyTable.R
\name{.prettifyTable}
\alias{.prettifyTable}
\title{test}
\usage{
.prettifyTable(
  table,
  filter,
  dom,
  show_buttons = FALSE,
  number_formatting = FALSE,
  color_highlighting = FALSE,
  hide_long_columns = FALSE,
  columns_percentage = NULL,
  columns_hide = NULL,
  download_file_name = NULL,
  page_length_default = 15,
  page_length_menu = c(15, 30, 50, 100)
)
}
\arguments{
\item{table}{test}

\item{groupA}{test}

\item{groupB}{test}

\item{mode}{test}

\item{percent}{test}
}
\value{
test
}
\description{
Function to prepare and format table.
}
