% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_randomlySubsetCells.R
\name{.randomlySubsetCells}
\alias{.randomlySubsetCells}
\title{test}
\usage{
.randomlySubsetCells(df, percentage)
}
\arguments{
\item{table}{test}

\item{groupA}{test}

\item{groupB}{test}

\item{mode}{test}

\item{percent}{test}
}
\value{
test
}
\description{
Randomly subset cells in data frame, if necessary.
}
