% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_GMT_file.R
\name{.read_GMT_file}
\alias{.read_GMT_file}
\title{Read GMT file.}
\usage{
.read_GMT_file(file)
}
\arguments{
\item{file}{Path to GMT file.}
}
\value{
Returns an object in the same format as from the GSA.read.gmt function (GSA
package) with gene sets, gene set names, and gene set descriptions stored in
lists.
}
\description{
This functions reads a (tab-delimited) GMT file which contains the gene set
name in the first column, the gene set description in the second column, and
the gene names in the following columns.
}
