% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_setRowOrder.R
\name{.setRowOrder}
\alias{.setRowOrder}
\title{test}
\usage{
.setRowOrder(df, order)
}
\arguments{
\item{table}{test}

\item{groupA}{test}

\item{groupB}{test}

\item{mode}{test}

\item{percent}{test}
}
\value{
test
}
\description{
Set order of rows in data frame.
}
