% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportFromSCE.R
\name{exportFromSCE}
\alias{exportFromSCE}
\title{Export SingleCellExperiment (SCE) object to Cerebro.}
\usage{
exportFromSCE(
  object,
  assay = "logcounts",
  file,
  experiment_name,
  organism,
  columns_groups,
  column_nUMI = "nUMI",
  column_nGene = "nGene",
  columns_cell_cycle = NULL,
  add_all_meta_data = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{object}{\code{SingleCellExperiment} (\code{SCE}) object.}

\item{assay}{Assay to pull expression values from; defaults to
\code{logcounts}. It is recommended to use sparse data (such as
log-transformed or raw counts) instead of dense data (such as the 'scaled'
slot) to avoid performance bottlenecks in the Cerebro interface.}

\item{file}{Where to save the output.}

\item{experiment_name}{Experiment name.}

\item{organism}{Organism, e.g. \code{hg} (human), \code{mm} (mouse), etc.}

\item{columns_groups}{Names of grouping variables in meta data
(\code{colData(object)}), e.g. \code{c("sample","cluster")}; at least one
must be provided; defaults to \code{NULL}.}

\item{column_nUMI}{Column in \code{colData(object)} that contains information
about number of transcripts per cell; defaults to \code{nUMI}.}

\item{column_nGene}{Column in \code{colData(object)} that contains
information about number of expressed genes per cell; defaults to
\code{nGene}.}

\item{columns_cell_cycle}{Names of columns in meta data
(\code{colData(object)}) that contain cell cycle information, e.g.
\code{c("Phase")}; defaults to \code{NULL}.}

\item{add_all_meta_data}{If set to \code{TRUE}, all further meta data columns
will be extracted as well.}

\item{verbose}{Set this to \code{TRUE} if you want additional log messages;
defaults to \code{FALSE}.}
}
\value{
No data returned.
}
\description{
This function allows to export a \code{SingleCellExperiment} (\code{SCE})
object to visualize in Cerebro.
}
\examples{
pbmc <- readRDS(system.file("extdata/v1.3/pbmc_SCE.rds",
  package = "cerebroApp"))
exportFromSCE(
  object = pbmc,
  file = 'pbmc_SCE.crb',
  experiment_name = 'PBMC',
  organism = 'hg',
  columns_groups = c('sample','cluster'),
  column_nUMI = 'nUMI',
  column_nGene = 'nGene',
  verbose = TRUE
)

}
