% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMarkerGenes.R
\name{getMarkerGenes}
\alias{getMarkerGenes}
\title{Get marker genes for every sample and cluster in Seurat object.}
\usage{
getMarkerGenes(
  object,
  assay = "RNA",
  organism = NULL,
  column_sample = "sample",
  column_cluster = "cluster",
  only_pos = TRUE,
  min_pct = 0.7,
  thresh_logFC = 0.25,
  thresh_p_val = 0.01,
  test = "wilcox",
  verbose = TRUE,
  ...
)
}
\arguments{
\item{object}{Seurat object.}

\item{assay}{Assay to pull counts from; defaults to 'RNA'. Only relevant in
Seurat v3.0 or higher since the concept of assays wasn't implemented before.}

\item{organism}{Organism information for pulling info about presence of
marker genes of cell surface; can be omitted if already saved in Seurat
object; defaults to NULL.}

\item{column_sample}{Column in object@meta.data that contains information
about sample; defaults to 'sample'.}

\item{column_cluster}{Column in object@meta.data that contains information
about cluster; defaults to 'cluster'.}

\item{only_pos}{Identify only over-expressed genes; defaults to TRUE.}

\item{min_pct}{Only keep genes that are expressed in at least n\% of current
group of cells, defaults to 0.70 (70\%).}

\item{thresh_logFC}{Only keep genes that show an average logFC of at least n;
defaults to 0.25.}

\item{thresh_p_val}{Threshold for p-value, defaults to 0.01.}

\item{test}{Statistical test used, defaults to 'wilcox' (Wilcoxon test).}

\item{verbose}{Print progress bar; defaults to TRUE.}

\item{...}{Further parameters can be passed to control
Seurat::FindAllMakers().}
}
\value{
Seurat object with marker gene results for samples and clusters
stored in object@misc$marker_genes.
}
\description{
This function gets marker genes for every sample and cluster of
the Seurat object.
}
\details{
Get marker genes for every sample and cluster in Seurat object.
}
\examples{
pbmc <- readRDS(system.file("extdata/v1.2/seurat_pbmc.rds",
  package = "cerebroApp"))
pbmc <- getMarkerGenes(
  object = pbmc,
  organism = 'hg',
  column_sample = 'sample',
  column_cluster = 'seurat_clusters',
  only_pos = TRUE,
  min_pct = 0.7,
  thresh_logFC = 0.25,
  thresh_p_val = 0.01,
  test = 'wilcox',
  verbose = TRUE
)
}
\keyword{Cerebro}
\keyword{Seurat}
\keyword{scRNAseq}
