% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMostExpressedGenes.R
\name{getMostExpressedGenes}
\alias{getMostExpressedGenes}
\title{Get most expressed genes for specified grouping variables in Seurat
object.}
\usage{
getMostExpressedGenes(object, assay = "RNA", groups = NULL)
}
\arguments{
\item{object}{Seurat object.}

\item{assay}{Assay to pull transcripts counts from; defaults to 'RNA'.}

\item{groups}{Grouping variables (columns) in object@meta.data for which most
expressed genes should be calculated; defaults to NULL.}
}
\value{
Seurat object with most expressed genes stored for every group level
of the specified groups stored in object@misc$most_expressed_genes.
}
\description{
This function calculates the most expressed genes for one or
multiple grouping variables in the meta data of the provided Seurat object.
}
\details{
Get most expressed genes for specified grouping variables in Seurat object.
}
\examples{
pbmc <- readRDS(system.file("extdata/v1.3/pbmc_seurat.rds",
  package = "cerebroApp"))
pbmc <- getMostExpressedGenes(
  object = pbmc,
  assay = 'RNA',
  groups = c('sample','seurat_clusters')
)
}
\keyword{Cerebro}
\keyword{Seurat}
\keyword{scRNAseq}
