% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launchApp.R
\name{launchApp}
\alias{launchApp}
\title{Launch Cerebro app.}
\usage{
launchApp(maxFileSize = 800)
}
\arguments{
\item{maxFileSize}{Maximum size of input file; defaults to 800 MB.}
}
\value{
Cerebro application object
}
\description{
Launch Cerebro app.
}
