% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launchCerebro.R
\name{launchCerebro}
\alias{launchCerebro}
\title{Launch Cerebro interface.}
\usage{
launchCerebro(version = "v1.2", maxFileSize = 800)
}
\arguments{
\item{version}{Which version of Cerebro to launch: "v1.0", "v1.1", "v1.2".
Defaults to "v1.2".}

\item{maxFileSize}{Maximum size of input file; defaults to 800 MB.}
}
\value{
Shiny application.
}
\description{
Launch Cerebro interface.
}
\details{
Launch Cerebro interface.
}
\examples{
if ( interactive() ) {
  launchCerebro(maxFileSize = 800)
}
}
\keyword{Cerebro}
\keyword{Seurat}
\keyword{scRNAseq}
