% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launchCerebroV1.0.R
\name{launchCerebroV1.0}
\alias{launchCerebroV1.0}
\title{Launch Cerebro v1.0}
\usage{
launchCerebroV1.0(maxFileSize = 800)
}
\arguments{
\item{maxFileSize}{Maximum size of input file; defaults to 800 MB.}
}
\value{
Shiny application.
}
\description{
Launch the Cerebro v1.0 Shiny application.
}
\details{
Launch Cerebro v1.0
}
\examples{
if ( interactive() ) {
  launchCerebroV1.0(maxFileSize = 800)
}
}
\keyword{Cerebro}
\keyword{Seurat}
\keyword{scRNAseq}
