% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launchCerebroV1.2.R
\name{launchCerebroV1.2}
\alias{launchCerebroV1.2}
\title{Launch Cerebro v1.2}
\usage{
launchCerebroV1.2(maxFileSize = 800)
}
\arguments{
\item{maxFileSize}{Maximum size of input file; defaults to 800 MB.}
}
\value{
Shiny application.
}
\description{
Launch the Cerebro v1.2 Shiny application.
}
\details{
Launch Cerebro v1.2
}
\examples{
if ( interactive() ) {
  launchCerebroV1.2(maxFileSize = 800)
}
}
\keyword{Cerebro}
\keyword{Seurat}
\keyword{scRNAseq}
