% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launchCerebroV1.3.R
\name{launchCerebroV1.3}
\alias{launchCerebroV1.3}
\title{Launch Cerebro v1.3}
\usage{
launchCerebroV1.3(
  mode = "open",
  maxFileSize = 800,
  crb_file_to_load = NULL,
  options = list()
)
}
\arguments{
\item{mode}{Cerebro can be ran in "open" or "closed" mode, allowing the user
to load their own data set ("open") or only show a pre-loaded data set
("closed", removes the "Load data" element); defaults to "open".}

\item{maxFileSize}{Maximum size of input file; defaults to 800 MB.}

\item{crb_file_to_load}{Path to .crb file to load on launch of Cerebro.
Useful when using/hosting Cerebro in "closed" mode. Defaults to NULL.}

\item{options}{...}
}
\value{
Shiny application.
}
\description{
Launch the Cerebro v1.3 Shiny application.
}
\details{
Launch Cerebro v1.3
}
\examples{
if ( interactive() ) {
  launchCerebroV1.3(
    mode = "open",
    maxFileSize = 800
  )
}
}
\keyword{Cerebro}
\keyword{Seurat}
\keyword{scRNAseq}
