% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/send_enrichr_query.r
\name{send_enrichr_query}
\alias{send_enrichr_query}
\title{Gene enrichment using Enrichr.}
\usage{
send_enrichr_query(genes, databases = NULL, URL_API = NULL)
}
\arguments{
\item{genes}{Gene names or dataframe of gene names in first column and a
score between 0 and 1 in the other.}

\item{databases}{Databases to search.}

\item{API_url}{URL to send requests to (Enrichr API).
See http://amp.pharm.mssm.edu/Enrichr/ for available databases.}
}
\value{
Returns a data frame of enrichment terms, p-values, ...
}
\description{
Gene enrichment using Enrichr.
}
\details{
Gene enrichment using Enrichr.
}
\author{
Wajid Jawaid, modified by Roman Hillje
}
