% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculatePercentGenes.R
\name{calculatePercentGenes}
\alias{calculatePercentGenes}
\title{Calculate percentage of transcripts of gene list.}
\usage{
calculatePercentGenes(object, genes)
}
\arguments{
\item{object}{Seurat object.}

\item{genes}{List(s) of genes.}
}
\value{
List of lists containing the percentages of expression for each
provided gene list.
}
\description{
Get percentage of transcripts of gene list compared to all
transcripts per cell.
}
\details{
Calculate percentage of transcripts of gene list.
}
\examples{
calculatePercentGenes(
  object = seurat,
  genes = list('example' = c('Fth1','Atf1'))
)
}
\keyword{Cerebro}
\keyword{Seurat}
\keyword{scRNAseq}
