% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportFromSeurat.R
\name{exportFromSeurat}
\alias{exportFromSeurat}
\title{Export Seurat object to Cerebro.}
\usage{
exportFromSeurat(
  object,
  assay = "RNA",
  slot = "data",
  file,
  experiment_name,
  organism,
  columns_groups,
  column_nUMI = "nUMI",
  column_nGene = "nGene",
  columns_cell_cycle = NULL,
  add_all_meta_data = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{object}{Seurat object.}

\item{assay}{Assay to pull expression values from; defaults to 'RNA'.}

\item{slot}{Slot to pull expression values from; defaults to 'data'. It is
recommended to use sparse data (such as log-transformed or raw counts)
instead of dense data (such as the 'scaled' slot) to avoid performance
bottlenecks in the Cerebro interface.}

\item{file}{Where to save the output.}

\item{experiment_name}{Experiment name.}

\item{organism}{Organism, e.g. hg (human), mm (mouse), etc.}

\item{columns_groups}{Names of grouping variables in meta data
(object@meta.data), e.g. c("sample","cluster"); at least one must be
provided; defaults to NULL.}

\item{column_nUMI}{Column in object@meta.data that contains information about
number of transcripts per cell; defaults to 'nUMI'.}

\item{column_nGene}{Column in object@meta.data that contains information
about number of expressed genes per cell; defaults to 'nGene'.}

\item{columns_cell_cycle}{Names of columns in meta data (object@meta.data)
that contain cell cycle information.e.g. c("Phase"); defaults to NULL.}

\item{add_all_meta_data}{If set to TRUE, all further meta data columns will
be extracted as well.}

\item{verbose}{Set this to TRUE if you want additional log messages; defaults
to FALSE.}
}
\value{
No data returned.
}
\description{
This function allows to export a Seurat object to visualize in
Cerebro.
}
\details{
Export Seurat object to Cerebro.
}
\examples{
pbmc <- readRDS(system.file("extdata/v1.2/seurat_pbmc.rds",
  package = "cerebroApp"))
exportFromSeurat(
  object = pbmc,
  file = 'pbmc_Seurat.crb',
  experiment_name = 'PBMC',
  organism = 'hg',
  columns_groups = c('sample','cluster'),
  column_nUMI = 'nCount_RNA',
  column_nGene = 'nFeature_RNA',
  columns_cell_cycle = c('Phase')
)
}
\keyword{Cerebro}
\keyword{Seurat}
\keyword{scRNAseq}
