% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMostExpressedGenes.R
\name{getMostExpressedGenes}
\alias{getMostExpressedGenes}
\title{Get most expressed genes for every sample and cluster in Seurat
object.}
\usage{
getMostExpressedGenes(object, column_sample = "sample",
  column_cluster = "cluster")
}
\arguments{
\item{object}{Seurat object.}

\item{column_sample}{Column in object@meta.data that contains information
about sample; defaults to 'sample'.}

\item{column_cluster}{Column in object@meta.data that contains information
about cluster; defaults to 'cluster'.}
}
\value{
Seurat object with most expressed genes stored for every sample and
cluster stored in object@misc$most_expressed_genes.
}
\description{
This function calculates the most expressed genes for every
sample and cluster of the Seurat object.
}
\details{
Get most expressed genes for every sample and cluster in Seurat object.
}
\examples{
seurat <- getMostExpressedGenes(
  object = seurat,
  column_sample = 'sample',
  column_cluster = 'cluster'
)
}
\keyword{Cerebro}
\keyword{Seurat}
\keyword{scRNAseq}
