% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launchCerebro.R
\name{launchCerebro}
\alias{launchCerebro}
\title{Launch Cerebro app.}
\usage{
launchCerebro(maxFileSize = 800)
}
\arguments{
\item{maxFileSize}{Maximum size of input file; defaults to 800 MB.}
}
\value{
Cerebro application object.
}
\description{
Launch Cerebro app.
}
\details{
Launch Cerebro app.
}
\examples{
launchApp(maxFileSize = 800)
}
\keyword{Cerebro}
\keyword{Seurat}
\keyword{scRNAseq}
