% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launchCerebroV1.2.R
\name{launchCerebroV1.2}
\alias{launchCerebroV1.2}
\title{Launch Cerebro v1.2}
\usage{
launchCerebroV1.2(maxFileSize = 800, ...)
}
\arguments{
\item{maxFileSize}{Maximum size of input file; defaults to \code{800}
(800 MB).}

\item{...}{Further parameters that are used by \code{shiny::runApp}, e.g.
\code{host} an \code{port}.}
}
\value{
Shiny application.
}
\description{
Launch the Cerebro v1.2 Shiny application.
}
\examples{
if ( interactive() ) {
  launchCerebroV1.2(maxFileSize = 800)
}

}
