% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launchCerebroV1.3.R
\name{launchCerebroV1.3}
\alias{launchCerebroV1.3}
\title{Launch Cerebro v1.3}
\usage{
launchCerebroV1.3(
  mode = "open",
  maxFileSize = 800,
  crb_file_to_load = NULL,
  ...
)
}
\arguments{
\item{mode}{Cerebro can be ran in \code{open} or \code{closed} mode, allowing
the user to load their own data set (\code{open}) or only show a pre-loaded
data set (\code{closed}, removes the "Load data" element); defaults to
\code{open}.}

\item{maxFileSize}{Maximum size of input file; defaults to \code{800}
(800 MB).}

\item{crb_file_to_load}{Path to \code{.crb} file to load on launch of
Cerebro. Useful when using/hosting Cerebro in \code{closed} mode. Defaults to
\code{NULL}.}

\item{...}{Further parameters that are used by \code{shiny::runApp}, e.g.
\code{host} an \code{port}.}
}
\value{
Shiny application.
}
\description{
Launch the Cerebro v1.3 Shiny application.
}
\examples{
if ( interactive() ) {
  launchCerebroV1.3(
    mode = "open",
    maxFileSize = 800
  )
}

}
