% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launchCerebroV1.3.R
\name{launchCerebroV1.3}
\alias{launchCerebroV1.3}
\title{Launch Cerebro v1.3}
\usage{
launchCerebroV1.3(
  mode = "open",
  maxFileSize = 800,
  crb_file_to_load = NULL,
  expression_matrix_h5 = NULL,
  welcome_message = NULL,
  projections_default_point_size = 5,
  projections_default_point_opacity = 1,
  projections_default_percentage_cells_to_show = 100,
  projections_show_hover_info = TRUE,
  ...
)
}
\arguments{
\item{mode}{Cerebro can be ran in \code{open} or \code{closed} mode, allowing
the user to load their own data set (\code{open}) or only show a pre-loaded
data set (\code{closed}, removes the "Load data" element); defaults to
\code{open}.}

\item{maxFileSize}{Maximum size of input file; defaults to \code{800}
(800 MB).}

\item{crb_file_to_load}{Path to \code{.crb} file to load on launch of
Cerebro. Useful when using/hosting Cerebro in \code{closed} mode. Defaults to
\code{NULL}.}

\item{expression_matrix_h5}{Path to \code{.h5} file containing an expression
matrix created with \code{HDF5Array::writeTENxMatrix()}, with genes as
columns and cells as rows, contrary to the conventional format of genes as
rows and cells as columns. This format greatly favors performance for
extracting expression values for a gene (column), rather than a cell (row),
which is the primary action in Cerebro. Importantly, the matrix should be 
stored with "expression" as group name (see parameters of the
\code{HDF5Array::writeTENxMatrix()} function). Saving the expression matrix
in \code{TENxMatrix} format has the benefit of a low memory footprint since
the expression values are directly read from disk. This is particularly
useful when working with very large data sets and/or when startup of the
Cerebro app is a priority (which is shorter because only the rest of the data
that needs to be loaded tends to be very small). By default, this value is
set to \code{NULL}, meaning that the expression matrix is expected to be part
of the \code{.crb} file.}

\item{welcome_message}{\code{string} with custom welcome message to display
in the "Load data" tab. Can contain HTML formatting, e.g.
\code{'<h3>Hi!</h3>'}. Defaults to \code{NULL}.}

\item{projections_default_point_size}{Default point size in projections. This
value can be changed in the UI; defaults to 5.}

\item{projections_default_point_opacity}{Default point opacity in
projections. This value can be changed in the UI; defaults to 1.0.}

\item{projections_default_percentage_cells_to_show}{Default percentage of
cells to show in projections. This value can be changed in the UI; defaults
to 100.}

\item{projections_show_hover_info}{Show hover infos in projections. This}

\item{...}{Further parameters that are used by \code{shiny::runApp}, e.g.
\code{host} or \code{port}.}
}
\value{
Shiny application.
}
\description{
Launch the Cerebro v1.3 Shiny application.
}
\examples{
if ( interactive() ) {
  launchCerebroV1.3(
    mode = "open",
    maxFileSize = 800
  )
}

}
